sinxi: Add support for Libre Computer Board ALL-H3-CC H5 ver.
The Libre Computer Board ALL-H3-CC from Libre Technology is a Raspberry
Pi B+ form factor single board computer based on the Allwinner H3 SoC.
The board has 1GB DDR3 SDRAM, provided by 4 2Gb chips. The mounting holes
and connectors are in the exact same position as on the Raspberry Pi B+.

This patch adds defconfig and dts files for the H5 variant of the same board.
The dts file simply includes the H3 variant since the board is the same.

The defconfig is based on the H3 variant and adapted to build for the H5.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>